Limit text two two-thirds width
> “Anything from 45 to 75 characters is widely regarded as a
> satisfactory length of line for a single-column page set in a serifed
> text face in a text size. The 66-character line (counting both letters
> and spaces) is widely regarded as ideal.

– http://webtypography.net/2.1.2

An easy way of achieving this when using the GOV.UK template is to limit
text to two thirds of the width of the main column. Doing so reduces the
first line of text on your page from 113 to 72 characters.

I also added a bit of spacing between the text and the charts so that
the leftmost charts don’t group too strongly with the paragraph above.